requirements.txt
Django==3.0.5
psycopg2-binary==2.8.5
django-braces==1.14.0
